refactor(app): split actions/list/detail into per-concern files (<500 each)
Final file-size violations: actions.go (887), list.go (1619), detail.go (978).
Pure reorganization, zero logic change — moved verbatim into same-package files:

  actions.go  -> actions_nav.go / actions_view.go / actions_list.go (+ thin actions.go)
  list.go     -> list_state.go (473) / list_body.go (425) / list_columns.go (425)
  detail.go   -> per-concern detail_*.go files

Largest internal/app file is now 473 lines (was 2543). Verified: ready-to-push
green, goldens byte-identical, lint 0, govulncheck clean.
